Updated def random.

+++ app/controllers/calculations_controller.rb
@@ -50,8 +50,8 @@ class CalculationsController < ApplicationController
   end
 
   def random
-    @min=params["min"].to_i
-    @max=params["max"].to_i
+    @min=params["min"].to_f
+    @max=params["max"].to_f
     @random_number=rand(@min..@max)
     render("calculations/random.html.erb")
   end
